refactor(k8s): move test-only cache/informer helpers into test files
